feat: Add zoom and pan to solution image
Add pinch-to-zoom and pan functionality to the solution image viewer.

This refactors the `SolutionViewerScreen` into a new, more capable `SolutionScreen.kt` file. The new screen allows users to zoom in up to 5x and pan around the solution image, ensuring no part of the image is panned out of view.

Other changes include:
*   Rename `BookDetailScreen` to `BookScreen`.
*   Rename `NavGraph.kt` to `NavHost.kt`.
*   Add a detailed `README.md`.