docs: Update license and README
This commit updates the license file and the README.

A disclaimer has been added to `LICENSE.md` to clarify that the application is unofficial and the developer is not responsible for the content, which is provided by an external service. The copyright year has also been updated to 2025.

Additionally, the "Screenshots" section in `README.md` has been moved to a more prominent position and formatted into a table for better presentation.